ADD REST calls for bootstrap and JSON structure
This commit will add rest calls to be called from client to bootstrap in
order to get information. The json classes will be used to convert Java
objects to json data and vice versa. The classes can be reused on client
side and should be it.

If the request somehow will fail the client will get a null respons and
not error codes this should be added in future