Add functionality for Json -> Pojo
Added classes to translate Bootstrap JSON to POJO. Now all JSON
data that bootstrap server is sending to client can be accessed easiley.

Pom file  update due to reasons that we used Jersey 1.xx
instead of the new version. Easier to parse information.

Maked sure that the Rest server will be possible to run on the client
also.